Implement CommonHttp module for HTTP requests
- Added CommonHttp struct with methods for GET and POST requests.
- Introduced HttpError enum for error handling.
- Created Response struct to encapsulate HTTP response data.
- Implemented utility functions for response parsing and header management.
- Integrated Cloudflare protection detection logic.

Enhance globals module with HTTP and FlareSolverr support

- Updated globals module to include CommonHttp and FlareSolverrManager.
- Refactored headless client to utilize new CommonHttp for network requests.
- Added utility functions for response handling in the globals context.

Introduce WASM bindings for HTTP and headless operations

- Created WASM bindings for HTTP requests and headless operations.
- Implemented FlareSolverr and headless client functionalities in WASM context.
- Added support for session management and element interaction in headless mode.

Clean up and modularize scraping utilities

- Moved scraping utilities to a dedicated module.
- Refactored HTML parsing and element selection logic for better maintainability.
- Enhanced error handling and response management across modules.

+++ scrapers/manga_dex/src/lib.rs
@@ -205,7 +205,7 @@ fn http_get(url: &str) -> Option<http::Response> {
 	let body = response.text().ok()?;
 
 	let response = http::Response {
-		status: status as u32,
+		status: status,
 		body,
 		headers: headers
 			.iter()
